requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pkg-resources==0.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
